requirements.txt
# Neural Navigator - Requirements
# ================================

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.21.0
Pillow>=9.0.0

# Visualization
matplotlib>=3.5.0

# Progress Bar
tqdm>=4.64.0

# Optional: GPU Support (CUDA 11.8)
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
